tidy: disable `google-runtime-float` globally instead of rewriting `long double`
Revert the `long double` -> `double` rewrites and the `NOLINT(google-runtime-float)`
suppressions in `shift10`, `CacheBase`, `HyperLogLogCounter`, `PrimeGenerator`,
`Defines.h`, `BloomFilterHash` and `InstrumentationManager`, and disable the check in
`.clang-tidy` instead.

`long double` is used deliberately in these places. In `shift10` the integer overloads
convert `UInt64`/`Int64` exactly and delay rounding until after the decimal shift, which
`readFloatTextFastImpl` relies on for correct `Float64` parsing near rounding boundaries;
rewriting it to `double` changed parsed values. A global disable is simpler than
per-site suppressions.

https://github.com/ClickHouse/ClickHouse/pull/108400

+++ src/Common/HyperLogLogCounter.h
@@ -98,7 +98,7 @@ struct IntermediateDenominator<UInt32, DenominatorType, denominator_mode>
 template <typename DenominatorType, DenominatorMode denominator_mode>
 struct IntermediateDenominator<UInt64, DenominatorType, denominator_mode>
 {
-    using Type = double;
+    using Type = long double;
 };
 
 template <typename HashValueType, typename DenominatorType>
@@ -179,10 +179,10 @@ public:
 
     DenominatorType get() const
     {
-        double val = rank_count[size - 1];
+        long double val = rank_count[size - 1];
         for (int i = size - 2; i >= 0; --i)
         {
-            val /= 2.0;
+            val /= 2.0L;
             val += rank_count[i];
         }
         return static_cast<DenominatorType>(val);
